Refactors names and imports to match Next conventions.
