+++ transaction-processor/src/app/components/WidthContainer/index.tsx
@@ -1,0 +1,3 @@
+import WidthContainer from "@/app/components/WidthContainer/WidthContainer";
+
+export { WidthContainer };
