+++ transaction-processor/src/app/components/button/index.tsx
@@ -1,0 +1,3 @@
+import Button from "@/app/components/Button/Button";
+
+export { Button };
