+++ transaction-processor/src/app/components/content/index.tsx
@@ -1,0 +1,3 @@
+import Content from '@/app/components/Content/Content';
+
+export { Content };
