+++ transaction-processor/src/app/components/input/index.tsx
@@ -1,0 +1,3 @@
+import Input from '@/app/components/Input/Input';
+
+export { Input };
