+++ transaction-processor/src/app/components/table/index.tsx
@@ -1,0 +1,3 @@
+import Table from "@/app/components/Table/Table";
+
+export { Table };
