+++ transaction-processor/src/app/features/Navbar/index.tsx
@@ -1,0 +1,3 @@
+import Navbar from '@/app/features/Navbar/components/Navbar';
+
+export { Navbar };
