+++ transaction-processor/src/app/features/accounts/index.tsx
@@ -1,0 +1,4 @@
+import Accounts from "@/app/features/accounts/components/Accounts";
+import AccountsByName from "./components/AccountsByName";
+
+export { Accounts, AccountsByName };
