+++ transaction-processor/src/app/features/all-transactions/index.tsx
@@ -1,0 +1,3 @@
+import AllTransactions from "@/app/features/all-transactions/components/AllTransactions";
+
+export { AllTransactions }
