+++ transaction-processor/src/app/features/bad-transactions/index.tsx
@@ -1,0 +1,3 @@
+import BadTransactions from "@/app/features/bad-transactions/components/BadTransactions";
+
+export { BadTransactions };
