+++ transaction-processor/src/app/features/collections/index.tsx
@@ -1,0 +1,3 @@
+import Collections from '@/app/features/collections/components/Collections';
+
+export { Collections };
