+++ transaction-processor/src/app/features/form/index.tsx
@@ -1,0 +1,3 @@
+import Form from "@/app/features/form/components/Form";
+
+export { Form }
